Overview:
- Bug fixes

- TrieMap mostly works now (except for remove)

- Added unit test capabilities

A    sspace/test
A    sspace/test/edu
A    sspace/test/edu/ucla
A    sspace/test/edu/ucla/sspace
A    sspace/test/edu/ucla/sspace/common

- Added a parallel directory structure for incorporating unit tests

A    sspace/test/edu/ucla/sspace/common/TrieMapTests.java

- Added an initial 20 unit tests for ensuring the functionality of the TrieMap class

AM   sspace/lib/junit-4.6.jar

- Added the junit test suite jar

M    sspace/src/edu/ucla/sspace/lsa/TfIdfTransformer.java
M    sspace/src/edu/ucla/sspace/lsa/LogEntropyTransformer.java
M    sspace/src/edu/ucla/sspace/tools/SparseMatrixConverter.java

- Added missing packages

M    sspace/src/edu/ucla/sspace/randomindexing/RandomIndexing.java

- Fixed bug where the random variance in the number of bits wasn't being
  calcuated properly.  This lead to index vectors always having one of two
  numbers of bits

M    sspace/src/edu/ucla/sspace/common/BoundedSortedMap.java

- Added serialVersionUID to fix lint warnings

M    sspace/src/edu/ucla/sspace/common/TrieMap.java

- Fixed bugs in put, get, size, isEmpty

- remove is still unimplemented

- Added collections views

- Added iterator support

M    sspace/src/edu/ucla/sspace/common/Similarity.java

- Fixed bug in covariance where integer division was being used instead of
  double divison

M    sspace/build.xml

- Added junit targets for building test classes and running tests

+++ trunk/sspace/src/edu/ucla/sspace/common/Similarity.java
@@ -82,8 +82,8 @@ public class Similarity {
 	    ySum += arr2[i];
 	}
 	
-	double xMean = xSum / (int)(arr1.length);
-	double yMean = ySum / (int)(arr1.length);
+	double xMean = xSum / (double)(arr1.length);
+	double yMean = ySum / (double)(arr1.length);
 	
 	double numerator = 0, xSqSum = 0, ySqSum = 0;
 	for (int i = 0; i < arr1.length; ++i) {
